+++ tests/escape_auto.py
@@ -1,6 +1,6 @@
 from docxtpl import *
 
-tpl = DocxTemplate("test_files/escape_tpl_auto.docx")
+tpl = DocxTemplate('templates/escape_tpl_auto.docx')
 
 context = {'myvar': R('"less than" must be escaped : <, this can be done with RichText() or R()'),
            'myescvar': 'It can be escaped with a "|e" jinja filter in the template too : < ',
@@ -10,4 +10,4 @@ context = {'myvar': R('"less than" must be escaped : <, this can be done with Ri
            }
 
 tpl.render(context, autoescape=True)
-tpl.save("test_files/escape_auto.docx")
+tpl.save('output/escape_auto.docx')