+++ tests/header_footer_entities.py
@@ -7,11 +7,11 @@ Created : 2015-03-12
 
 from docxtpl import DocxTemplate
 
-tpl=DocxTemplate('test_files/header_footer_entities_tpl.docx')
+tpl=DocxTemplate('templates/header_footer_entities_tpl.docx')
 
 context = {
     'title' : 'Header and footer test',
 }
 
 tpl.render(context)
-tpl.save('test_files/header_footer_entities.docx')
+tpl.save('output/header_footer_entities.docx')