+++ tests/header_footer_image.py
@@ -7,13 +7,13 @@ Created : 2017-09-03
 
 from docxtpl import DocxTemplate
 
-DEST_FILE = 'test_files/header_footer_image.docx'
+DEST_FILE = 'output/header_footer_image.docx'
 
-tpl=DocxTemplate('test_files/header_footer_image_tpl.docx')
+tpl=DocxTemplate('templates/header_footer_image_tpl.docx')
 
 context = {
     'mycompany' : 'The World Wide company',
 }
-tpl.replace_media('test_files/dummy_pic_for_header.png','test_files/python.png')
+tpl.replace_media('templates/dummy_pic_for_header.png','templates/python.png')
 tpl.render(context)
 tpl.save(DEST_FILE)