+++ tests/header_footer_utf8.py
@@ -9,7 +9,7 @@ Edited : 2016-07-19 by Eric Lapouyade
 
 from docxtpl import DocxTemplate
 
-tpl=DocxTemplate('test_files/header_footer_tpl_utf8.docx')
+tpl=DocxTemplate('templates/header_footer_tpl_utf8.docx')
 
 sd = tpl.new_subdoc()
 p = sd.add_paragraph(u'This is a sub-document to check it does not break header and footer with utf-8 characters inside the template .docx')
@@ -22,4 +22,4 @@ context = {
 }
 
 tpl.render(context)
-tpl.save('test_files/header_footer_utf8.docx')
+tpl.save('output/header_footer_utf8.docx')