+++ tests/replace_picture.py
@@ -7,12 +7,12 @@ Created : 2017-09-03
 
 from docxtpl import DocxTemplate
 
-DEST_FILE = 'test_files/replace_picture.docx'
+DEST_FILE = 'output/replace_picture.docx'
 
-tpl=DocxTemplate('test_files/replace_picture_tpl.docx')
+tpl=DocxTemplate('templates/replace_picture_tpl.docx')
 
 context = {}
 
-tpl.replace_pic('python_logo.png','test_files/python.png')
+tpl.replace_pic('python_logo.png','templates/python.png')
 tpl.render(context)
 tpl.save(DEST_FILE)