+++ tests/template_error.py
@@ -6,7 +6,7 @@ six.print_('=' * 80)
 six.print_("Generating template error for testing (so it is safe to ignore) :")
 six.print_('.' * 80)
 try:
-    tpl = DocxTemplate('test_files/template_error_tpl.docx')
+    tpl = DocxTemplate('templates/template_error_tpl.docx')
     tpl.render({
         'test_variable' : 'test variable value'
     })
@@ -16,7 +16,7 @@ except TemplateError as the_error:
         six.print_("Context:")
         for line in the_error.docx_context:
             six.print_(line)
-tpl.save('test_files/template_error.docx')
+tpl.save('output/template_error.docx')
 six.print_('.' * 80)
 six.print_(" End of TemplateError Test ")
 six.print_('=' * 80)